Make newlines in static trigger newlines in generated source:
Earlier, [:static, "foo\nbar"] would be generated into
_buf << "foo\nbar".  This makes perfectly sense, but most
of the time a newline in the static actually means there's
a newline in the template source.  In HEAD^ you had to use
[:static, "foo\nbar"], [:newline] in order to fix that.

With this commit, it will now be generated into: _buf << "foo
bar", and therefore we don't need the :newline anymore.

This also means we'll have to introduce another type which
behaves like static did before.