Separate attribute sorting and merging responsibilities
Just as one might want to merge attributes but not sort them (see
b45c61), one might want to sort attributes but not merge them.
They are separate concerns, and therefore should be separate filters.
This change permits the use of AttributeSorter independently from
AttributeMerger.

Backward compatibility requires preserving the existing sorting
behavior of AttributeMerger, but in a future release it should be
fully separated, and users should explicitly use AttributeSorter
in the filter chain if they desire.