Reduce string allocations
We can modify the html string in place using `gsub` instead of allocating a new string.

This may be a breaking change, but the tests still pass...

On my app I'm seeing 1843 fewer objects and 68,992 bytes saved per request.


For really minor gains, we can check if an element is nil and return a frozen empty string. This gives us a savings of 95 objects.

It looks like there are possibly more savings here, but they'll be a bit harder to get at

```
      1776  "&|\"|'|<|>"
      1776  /Users/richardschneeman/Documents/projects/temple/lib/temple/utils.rb:51
```

It looks like gsub! and gub both allocates strings internally when using hash arguments. It might be worth looking into how loofah does sanitization, or maybe looking into alternate ways to do the sanitization.