Always escape quotes in html attribute values
By having quotes inside html attribute values, the generated html code
can get wrong. For example

```
[:html, :tag, 'div',
  [:html, :attrs, [
    [:html, :attr, 'data-something', [:static, '"test']]
  ], [:content]
]
```

compiles to

```html
<div data-something=""test"></div>
```

Notice the `"` closes `data-something` which obviously is not intended.
To fix this, all quotes no always get escaped, so generated code
becomes:

```html
<div data-something="&quot;test"></div>
```

`"` is escaped to `&quot;` and `'` is escaped to `&#34`, so both
delimiters can be used.